cake: correct incidence angle computation in case of a ray arriving from above
Before this fix, cake was reporting incorrect values for incindence
angles, when a ray was arriving downward at the receiver (burried
receiver). The case of a ray arriving at the receiver from below and the
take-off angle computations in both cases were not affected by the bug.
The buggy version produced values of

   erroneous_angle = 270 - correct_angle