gf: refactor time handling
Changed how origin times are handled in seismogram synthesis. Before,
stacking was done for an origin time of zero and ready seismograms where
shifted to the correct origin time in post-processing. Now the
discretized point sources have their true time set. Only just before and
after the call to the C functions, time shifting is done. This
simplifies the code in various places but disables the optimization to
do only one stack for sources differing only in their origin time.

The former approach was incompatible with tmin-tmax limiting on the
receivers.

- removed method Source.get_timeshift()